+++ esp-rtos/src/lib.rs
@@ -1,6 +1,13 @@
-//! This crate allows using esp-radio on top of esp-hal, without any other OS.
+#![cfg_attr(
+    all(docsrs, not(not_really_docsrs)),
+    doc = "<div style='padding:30px;background:#810;color:#fff;text-align:center;'><p>You might want to <a href='https://docs.espressif.com/projects/rust/'>browse the <code>esp-hal</code> documentation on the esp-rs website</a> instead.</p><p>The documentation here on <a href='https://docs.rs'>docs.rs</a> is built for a single chip only (ESP32-C6, in particular), while on the esp-rs website you can select your exact chip from the list of supported devices. Available peripherals and their APIs change depending on the chip.</p></div>\n\n<br/>\n\n"
+)]
+//! This crate provides RTOS functionality for `esp-radio`, and provides executors to enable
+//! running `async` code.
 //!
-//! This crate requires an esp-hal timer to operate.
+//! ## Setup
+//!
+//! This crate requires an esp-hal timer to operate, and needs to be started like so:
 //!
 //! ```rust, no_run
 //! use esp_hal::timer::timg::TimerGroup;
@@ -51,10 +58,17 @@ esp_rtos::start_second_core(
 //! // let esp_radio_controller = esp_radio::init().unwrap();
 //! # }
 //! ```
+//! 
+//! To write `async` code, enable the `embassy` feature, and mark the main function with `#[esp_rtos::main]`.
+//! Note that, to create async tasks, you will need the `task` macro from the `embassy-executor` crate. Do
+//! NOT enable any of the `arch-*` features on `embassy-executor`.
+//!
 //! ## Feature Flags
 #![doc = document_features::document_features!()]
 #![no_std]
 #![cfg_attr(xtensa, feature(asm_experimental_arch))]
+#![cfg_attr(docsrs, feature(doc_cfg))]
+#![deny(missing_docs)]
 
 #[cfg(feature = "alloc")]
 extern crate alloc;
@@ -72,6 +86,7 @@ mod timer;
 mod wait_queue;
 
 #[cfg(feature = "embassy")]
+#[cfg_attr(docsrs, doc(cfg(feature = "embassy")))]
 pub mod embassy;
 
 use core::mem::MaybeUninit;
@@ -91,6 +106,7 @@ use esp_hal::{
     system::{CpuControl, Stack},
     time::{Duration, Instant},
 };
+#[cfg_attr(docsrs, doc(cfg(feature = "embassy")))]
 pub use macros::rtos_main as main;
 pub(crate) use scheduler::SCHEDULER;
 pub use task::CurrentThreadHandle;
@@ -244,6 +260,35 @@ pub fn start_second_core<const STACK_SIZE: usize>(
     stack: &'static mut Stack<STACK_SIZE>,
     func: impl FnOnce() + Send + 'static,
 ) {
+    start_second_core_with_stack_guard_offset::<STACK_SIZE>(
+        cpu_control,
+        #[cfg(xtensa)]
+        int0,
+        int1,
+        stack,
+        Some(esp_config::esp_config_int!(
+            usize,
+            "ESP_HAL_CONFIG_STACK_GUARD_OFFSET"
+        )),
+        func,
+    );
+}
+
+/// Starts the scheduler on the second CPU core.
+///
+/// Note that the scheduler must be started first, before starting the second core.
+///
+/// The supplied stack and function will be used as the main thread of the second core. The thread
+/// will be pinned to the second core.
+#[cfg(multi_core)]
+pub fn start_second_core_with_stack_guard_offset<const STACK_SIZE: usize>(
+    cpu_control: CPU_CTRL,
+    #[cfg(xtensa)] int0: SoftwareInterrupt<'static, 0>,
+    int1: SoftwareInterrupt<'static, 1>,
+    stack: &'static mut Stack<STACK_SIZE>,
+    stack_guard_offset: Option<usize>,
+    func: impl FnOnce() + Send + 'static,
+) {
     trace!("Starting scheduler for the second core");
 
     #[cfg(xtensa)]
@@ -262,7 +307,7 @@ pub fn start_second_core<const STACK_SIZE: usize>(
 
     let mut cpu_control = CpuControl::new(cpu_control);
     let guard = cpu_control
-        .start_app_core(stack, move || {
+        .start_app_core_with_stack_guard_offset(stack, stack_guard_offset, move || {
             trace!("Second core running");
             task::setup_smp(int1);
             SCHEDULER.with(move |scheduler| {